Moved SID/MID signing to NetworkBackend Added JSONConfiguration
